Create, edit and delete item by form

+++ app6.py
@@ -1,5 +1,6 @@
 # 第 3 章：模板
 from flask import Flask, render_template
+from flask import request, url_for, redirect, flash
 from flask_sqlalchemy import SQLAlchemy
 
 app = Flask(__name__)
@@ -27,7 +28,9 @@ class Movie(db.Model):  # 表名将会是 movie
     title = db.Column(db.String(60))  # 电影标题
     year = db.Column(db.String(4))  # 电影年份
 
-def fake_data(refresh = False):
+def fake_data(run = True, refresh = False):
+    if not run:
+        return 0
     if refresh:
         db.drop_all()
         db.create_all()
@@ -66,14 +69,62 @@ def inject_user():
 def page_not_found(e):
     return render_template('404.html'), 404
 
-
-@app.route('/')
+app.config['SECRET_KEY'] = 'dev'  # 等同于 app.secret_key = 'dev'
+@app.route('/', methods=['GET', 'POST'])
 def index():
     user = User.query.first()# 利用外键找到user的movie单
+    # 这里先认为只有一个用户
+    if request.method == 'POST':  # 判断是否是 POST 请求
+        # 获取表单数据
+        title = request.form.get('title')  # 传入表单对应输入字段的 name 值
+        year = request.form.get('year')
+        # 验证数据
+        if not title or not year or len(year) > 4 or len(title) > 60:
+            flash('Invalid input.')  # 显示错误提示
+            return redirect(url_for('index'))  # 重定向回主页
+        # 保存表单数据到数据库
+        movie = Movie(title=title, year=year, user_id = user.id)  # 创建记录
+        db.session.add(movie)  # 添加到数据库会话
+        db.session.commit()  # 提交数据库会话
+        flash('Item created.')  # 显示成功创建的提示
+        return redirect(url_for('index'))  # 重定向回主页
+    
     movies = Movie.query.filter_by(user_id = user.id).all()
     return render_template('index.html', movies=movies)
 
+# 编辑电影条目
+@app.route('/movie/edit/<int:movie_id>', methods=['GET', 'POST'])
+def edit(movie_id):
+    movie = Movie.query.get_or_404(movie_id)
+
+    if request.method == 'POST':  # 处理编辑表单的提交请求
+        title = request.form['title']
+        year = request.form['year']
+
+        if not title or not year or len(year) != 4 or len(title) > 60:
+            flash('Invalid input.')
+            return redirect(url_for('edit', movie_id=movie_id))  # 重定向回对应的编辑页面
+
+        movie.title = title  # 更新标题
+        movie.year = year  # 更新年份
+        db.session.commit()  # 提交数据库会话
+        flash('Item updated.')
+        return redirect(url_for('index'))  # 重定向回主页
+
+    return render_template('edit.html', movie=movie)  # 传入被编辑的电影记录
+
+# 删除
+@app.route('/movie/delete/<int:movie_id>', methods=['POST'])  # 限定只接受 POST 请求
+def delete(movie_id):
+    movie = Movie.query.get_or_404(movie_id)  # 获取电影记录
+    db.session.delete(movie)  # 删除对应的记录
+    db.session.commit()  # 提交数据库会话
+    flash('Item deleted.')
+    return redirect(url_for('index'))  # 重定向回主页
+
+
 if __name__ == "__main__":
-    with app.app_context():
-        fake_data(refresh=True)
+    # with app.app_context():
+    #     fake_data(refresh=True)
+    fake_data(run=False)
     app.run()
